pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-order-lookup"
version = "1.0.0"
description = "Desktop application for accessing Visual SQL Server database order history"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Spare Parts Department"}
]
keywords = ["visual", "database", "orders", "pyqt6", "desktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "PyQt6==6.6.0",
    "pyodbc==5.0.1",
    "python-dotenv==1.0.0",
    "Jinja2==3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-qt==4.2.0",
    "pytest-mock==3.12.0",
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "pyinstaller==6.3.0",
]

[project.scripts]
visual-order-lookup = "visual_order_lookup.main:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
]
